Update FlowStatus module to use the 'yarn.resourcemanager.webapp.address' property for jt_url
Originally, we were using 'mapred.job.tracker', which gave us the jobtracker URL. That property is deprecated. We are now using 'yarn.resourcemanager.webapp.address' to determine the resourcemanager URL. For backwards compatibility, we fall back on 'mapred.job.tracker' should we be unable to find 'yarn.resourcemanager.webapp.address'.