Correcting the "cluster" link
The current path `?cluster=foo` will always append to the current URI path, which means that from a flowgraph page it just append that argument, rather than showing the list of jobs filtered to that cluster. By specifying the leading slash `/?cluster=foo`, the URL will be correctly set on all pages.